refactor: move adminPage/actorCastingPage to pages folder
