+++ apps/web/src/modules/job/list/pages/ActorCastingPage/components/NotiCardContainer/components/NotiCard/index.tsx
