+++ apps/web/src/modules/job/list/pages/ActorCastingPage/components/NotiCardContainer/styled.ts
