+++ apps/web/src/modules/job/list/pages/ActorCastingPage/index.tsx
@@ -4,13 +4,13 @@ import { Button, CircularProgress, Typography } from '@mui/material'
 import { useUser } from 'common/context/UserContext'
 import withGuard from 'common/hoc/withGuard'
 import useNavbarSearch from 'common/hooks/useNavbarSearch'
+import FilterContainer from 'modules/job/list/components/FilterContainer'
+import FilterMobileContainer from 'modules/job/list/components/FilterMobileContainer'
+import JobCardContainer from 'modules/job/list/components/JobCardContainer'
+import SearchBox from 'modules/job/list/components/SearchBox'
 import React, { useCallback } from 'react'
 import InfiniteScroll from 'react-infinite-scroller'
 
-import FilterContainer from '../FilterContainer'
-import FilterMobileContainer from '../FilterMobileContainer'
-import JobCardContainer from '../JobCardContainer'
-import SearchBox from '../SearchBox'
 import NotiCardContainer from './components/NotiCardContainer'
 import useJobListData from './hooks/useJobListData'
 import { notiHolder } from './placeholder'
