+++ apps/web/src/modules/job/list/pages/AdminPage/styled.ts
